PROGRAM COMPLETE - Taxi Program can generate, update, and save/load the q table. Current Task List (if I ever return): - command line parameters can be used to replace the Q table related global parameters located in main.py. This will allow the user to save/load the q table and have the option to skip training time without having to make code changes. - Taxi.reduce_exploration() is too static and could reference the number of training episodes to maximize exploration time for the amount of episodes given.
Revelation:
Taxi-v3 is not a game about picking up and dropping off passengers, it is a game which has a set path of rewards. If the player stays on the golden path (indicated by the reward provided after each step), their total score will be high. However, if the player takes an alternate route, even if they achieve their goal in the same number of steps, the resulting score is much lower. This is what the Q tables are meant to pick up on.